# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/calendar_branches:
    get:
      tags:
        - CalendarBranches
      summary: List CalendarBranches
      description: Get a paginated list of CalendarBranches with optional filtering and sorting
      operationId: listCalendarBranches
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarBranchList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarBranches
      summary: Create CalendarBranch
      description: Create a new CalendarBranch
      operationId: createCalendarBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarBranchRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarBranch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/calendar_departments:
    get:
      tags:
        - CalendarDepartments
      summary: List CalendarDepartments
      description: Get a paginated list of CalendarDepartments with optional filtering and sorting
      operationId: listCalendarDepartments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarDepartmentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarDepartments
      summary: Create CalendarDepartment
      description: Create a new CalendarDepartment
      operationId: createCalendarDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarDepartmentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarDepartment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    CalendarBranch:
      type: object
      required:
        - id
        - calendar_id
        - company_id
        - branch_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        calendar_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCalendarBranchRequest:
      type: object
      required:
        - calendar_id
        - company_id
        - branch_id
        - metadata
      properties:
        calendar_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CalendarBranchList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarBranch'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CalendarDepartment:
      type: object
      required:
        - id
        - calendar_id
        - department_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        calendar_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCalendarDepartmentRequest:
      type: object
      required:
        - calendar_id
        - department_id
        - metadata
      properties:
        calendar_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CalendarDepartmentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarDepartment'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []